+++ src/main/java/com/zerobase/foodlier/module/member/member/mail/service/MailService.java
@@ -1,4 +1,4 @@
-package com.zerobase.foodlier.global.member.mail.service;
+package com.zerobase.foodlier.module.member.member.mail.service;
 
 public interface MailService {
     void sendMail(String email, String verificationCode);
