+++ src/main/java/com/zerobase/foodlier/module/member/member/mail/service/VerificationCodeService.java
@@ -1,4 +1,4 @@
-package com.zerobase.foodlier.global.member.mail.service;
+package com.zerobase.foodlier.module.member.member.mail.service;
 
 public interface VerificationCodeService {
     String createAuthenticationCode();
